+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+* Mangle Syntax `Infinity` to `1/0`
+
+	MinifySyntax will now mangle the `Infinty` keywoard to `1/0` which is less in bytes.
+
 ## 0.12.9
 
 * Allow `this` with `--define` ([#1361](https://github.com/evanw/esbuild/issues/1361))